Moved userIndicator to containers
