+++ src/components/Header.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 
-import UserIndicator from './UserIndicator';
+import UserIndicator from '../containers/UserIndicator';
 
 class Header extends React.Component {
