+++ src/containers/UserIndicator.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { NavLink } from 'react-router-dom';
-import { logout } from '../actions/app';
+import { logout, subnavStatus } from '../actions/app';
 
 class UserIndicator extends React.Component {
   constructor(props) {
@@ -10,14 +10,11 @@ class UserIndicator extends React.Component {
     this.handleClick = this.handleClick.bind(this);
     this.toggleSubNav = this.toggleSubNav.bind(this);
 
-    this.state = {
-      isActive: false
-    }
   }
 
   render() {
     let classes = "user-indicator__menu sub-nav";
-    if (this.state.isActive) {
+    if (this.props.subnavState) {
       classes += " active";
     }
 
@@ -33,8 +30,13 @@ class UserIndicator extends React.Component {
   }
 
   toggleSubNav(event) {
-    const STATE = this.state.isActive ? false : true;
-    this.setState({isActive: STATE});
+    event.preventDefault();
+
+    const STATE = this.props.subnavState ? false : true;
+
+    let reduxAction = subnavStatus(STATE);
+    this.setState({subnavState: STATE});
+    this.props.dispatch(reduxAction);
   }
 
   handleClick(event) {
@@ -48,5 +50,6 @@ class UserIndicator extends React.Component {
 export default connect(function(state){
   return {
     location: state.router.location,
+    subnavState: state.app.subnavState
   };
 })(UserIndicator);
